pyproject.toml
[tool.poetry]
name = "ai_convo"
version = "0.1.0"
description = "Georgian GenAI Bootcamp - 1 - 2023"
authors = ["rodrigo-georgian"]

packages = [
    { include = "ai_convo", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
jupyter = "^1.0.0"
openai = "^0.27.7"
jupyterlab = "^4.0.1"
google-cloud-aiplatform = "^1.25.0"
pandas = "^1.3.5"
seaborn = "^0.12.2"
langchain = "^0.0.200"
fiddler-auditor = "^0.0.1"
torch = "2.0.1"
faiss-cpu = "^1.7.4"
tiktoken = "^0.4.0"
datasets = "2.12.0"
evaluate = "0.4.0"
numpy = "1.24.3"
tqdm = "4.65.0"
transformers = "4.30.0"
ipykernel = "^6.23.2"
ipywidgets = "^8.0.6"
peft = {git = "https://github.com/huggingface/peft"}
privateai-client = "^1.2.0"

[tool.poetry.dev-dependencies]
pytest = "^7.0.0"
pre-commit = "^2.15.0"
black = "^22.0.0"
mypy = "^0.971"
isort = "^5.9.3"
flake8 = "^5.0.0"
pytest-cov = "^3.0.0"
coverage = {version = "^6.0.1", extras = ["toml"]}

[tool.semantic_release]
branch = "main"
version_toml = "pyproject.toml:tool.poetry.version"
version_variable = "src/genai_bootcamp/__init__.py:__version__"
build_command = "pip install poetry && poetry build"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.run]
branch = true

[tool.coverage.report]
show_missing = true

[tool.black]
line-length = 120
target-version = ['py39']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''

[tool.isort]
profile = "black"
